+++ src/components/index.js
@@ -1,4 +1,5 @@
 // Keep in alphabetical order
 export { default as Board } from './Board/Board';
-export { default as Square } from './Square/Square';
+export { default as Game } from './Game/Game';
 export { default as Piece } from './Piece/Piece';
+export { default as Square } from './Square/Square';