+++ CHANGELOG.md
@@ -18,10 +18,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Security
 ---
 
+## [0.0.11] 2018-08-24
+
+### Added
+-
+
+## [0.0.11] 2018-08-23
+
+### Added
+- <Game/> component
+
 ## [0.0.1] 2018-08-22
 
 ### Added
 - Initiate project with create-react-app
 - Content for changelog, readme, license files
 - Storybook and ESLint
-- First components (Square, Board)
+- First components: <Square/>, <Board/>, <Piece/>
+- PNG images for the pieces